Refactor LocationView to improve clarity and fix a couple of bugs.
- Move DefaultLocationsList back into LocationView, instead of making it its own component
- Fixed interactions between the default search locations and locations set in the search bar
- Moved the search bar text into a @State in LocationView
